chore(http): split ErrorReponse from OkResponse
fixes #12899

BREAKING CHANGE
