+++ engine/tests/phpunit/unit/Elgg/Http/ErrorResponseTest.php
@@ -6,7 +6,7 @@ namespace Elgg\Http;
  * @group HttpService
  * @group UnitTests
  */
-class ErrorResponseTest extends OkResponseUnitTest {
+class ErrorResponseUnitTest extends ResponseUnitTest {
 
 	public function up() {
 		$this->class = ErrorResponse::class;
@@ -25,12 +25,6 @@ class ErrorResponseTest extends OkResponseUnitTest {
 		$this->assertEquals([], $response->getHeaders());
 	}
 
-	public function testConstructWithInvalidStatusCode() {
-		$test_class = $this->class;
-		$response = new $test_class('foo', 9999);
-		$this->assertEquals(ELGG_HTTP_INTERNAL_SERVER_ERROR, $response->getStatusCode());
-	}
-
 	public function testCanConstructWithArguments() {
 		$error = 'foo';
 		$status_code = ELGG_HTTP_NOT_FOUND;
@@ -44,6 +38,12 @@ class ErrorResponseTest extends OkResponseUnitTest {
 		$this->assertEquals($forward_url, $response->getForwardURL());
 		$this->assertEquals([], $response->getHeaders());
 	}
+	
+	public function testConstructWithInvalidStatusCode() {
+		$test_class = $this->class;
+		$response = new $test_class('foo', 9999);
+		$this->assertEquals(ELGG_HTTP_INTERNAL_SERVER_ERROR, $response->getStatusCode());
+	}
 
-	// Remaining tests are identical to OkResponseUnitTest
+	// Remaining tests are identical to ResponseUnitTest
 }
